feat(core): Extract StreamReference (immutable typed reference to file)
Every time `Var.stream()` is invoked, an `InputStream` related to the _current_ file value is returned. `StreamReference`, on the other hand, is an _immutable reference_ to such stream, allowing to predictably obtain an `InputStream` pointing always to the same location. It also provides typed access to specific information related to where it originates (resource/asset/file/memory).